Move userDetail out of Moderation route
