+++ client/coral-admin/src/components/ButtonCopyToClipboard.js
