+++ client/coral-admin/src/components/FlagBox.js
