+++ client/coral-admin/src/components/UserDetailComment.js
